+++ data/types/device/status/test/status.go
@@ -15,6 +15,7 @@ func NewStatus() *status.Status {
 	datum.Device = *testDataTypesDevice.NewDevice()
 	datum.SubType = "status"
 	datum.Duration = pointer.Int(test.RandomIntFromRange(status.DurationMinimum, math.MaxInt32))
+	datum.DurationExpected = pointer.Int(test.RandomIntFromRange(*datum.Duration, math.MaxInt32))
 	datum.Name = pointer.String(test.RandomStringFromArray(status.Names()))
 	datum.Reason = testData.NewBlob()
 	return datum
@@ -27,6 +28,7 @@ func CloneStatus(datum *status.Status) *status.Status {
 	clone := status.New()
 	clone.Device = *testDataTypesDevice.CloneDevice(&datum.Device)
 	clone.Duration = test.CloneInt(datum.Duration)
+	clone.DurationExpected = test.CloneInt(datum.DurationExpected)
 	clone.Name = test.CloneString(datum.Name)
 	clone.Reason = testData.CloneBlob(datum.Reason)
 	return clone